requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
cryptography
requests==2.31.0
python-dateutil==2.8.2
azure-storage-file-share==12.5.0
snowflake-connector-python==3.0.2
cffi==1.14.6